refactor(sdk-logs): extract addAttribute helper
