+++ experimental/packages/sdk-logs/src/LogRecordImpl.ts
@@ -21,8 +21,8 @@ import {
 } from '@opentelemetry/semantic-conventions';
 import type { ReadableLogRecord } from './export/ReadableLogRecord';
 import type { LogRecordLimits } from './types';
-import { isLogAttributeValue } from './utils/validation';
 import type { LoggerProviderSharedState } from './internal/LoggerProviderSharedState';
+import { addAttribute, AddAttributeDecision } from './utils/attributes';
 
 export class LogRecordImpl implements ReadableLogRecord {
   readonly hrTime: api.HrTime;
@@ -132,33 +132,25 @@ export class LogRecordImpl implements ReadableLogRecord {
     if (this._isLogRecordReadonly()) {
       return this;
     }
-    if (key.length === 0) {
-      api.diag.warn(`Invalid attribute key: ${key}`);
-      return this;
-    }
-    if (!isLogAttributeValue(value)) {
-      api.diag.warn(`Invalid attribute value set for key: ${key}`);
-      return this;
-    }
-    const isNewKey = !Object.prototype.hasOwnProperty.call(
+
+    const decision = addAttribute(
       this.attributes,
-      key
+      this._logRecordLimits,
+      this._attributesCount,
+      key,
+      value
     );
-    if (
-      isNewKey &&
-      this._attributesCount >= this._logRecordLimits.attributeCountLimit
-    ) {
+
+    if (decision === AddAttributeDecision.DROP_LIMIT_REACHED) {
       this._droppedAttributesCount++;
-      // Only warn once per LogRecord to avoid log spam
       if (this._droppedAttributesCount === 1) {
+        // Only warn once per LogRecord to avoid log spam
         api.diag.warn('Dropping extra attributes.');
       }
-      return this;
-    }
-    this.attributes[key] = this._truncateToSize(value);
-    if (isNewKey) {
+    } else if (decision === AddAttributeDecision.ADD_NEW) {
       this._attributesCount++;
     }
+
     return this;
   }
 
@@ -196,48 +188,6 @@ export class LogRecordImpl implements ReadableLogRecord {
    */
   _makeReadonly() {
     this._isReadonly = true;
-  }
-
-  private _truncateToSize(value: AnyValue): AnyValue {
-    const limit = this._logRecordLimits.attributeValueLengthLimit;
-    // Check limit
-    if (limit <= 0) {
-      // Negative values are invalid, so do not truncate
-      api.diag.warn(`Attribute value limit must be positive, got ${limit}`);
-      return value;
-    }
-
-    // null/undefined - no truncation needed
-    if (value == null) {
-      return value;
-    }
-
-    // String
-    if (typeof value === 'string') {
-      return this._truncateToLimitUtil(value, limit);
-    }
-
-    // Byte arrays - no truncation needed
-    if (value instanceof Uint8Array) {
-      return value;
-    }
-
-    // Arrays (can contain any AnyValue types)
-    if (Array.isArray(value)) {
-      return value.map(val => this._truncateToSize(val));
-    }
-
-    // Objects/Maps - recursively truncate nested values
-    if (typeof value === 'object') {
-      const truncatedObj: Record<string, AnyValue> = {};
-      for (const [k, v] of Object.entries(value as Record<string, AnyValue>)) {
-        truncatedObj[k] = this._truncateToSize(v);
-      }
-      return truncatedObj;
-    }
-
-    // Other types (number, boolean), no need to apply value length limit
-    return value;
   }
 
   private _setException(exception: unknown): void {
@@ -286,13 +236,6 @@ export class LogRecordImpl implements ReadableLogRecord {
     if (!hasMinimumAttributes) {
       api.diag.warn(`Failed to record an exception ${exception}`);
     }
-  }
-
-  private _truncateToLimitUtil(value: string, limit: number): string {
-    if (value.length <= limit) {
-      return value;
-    }
-    return value.substring(0, limit);
   }
 
   private _isLogRecordReadonly(): boolean {
