+++ experimental/packages/sdk-logs/test/common/utils/attributes.test.ts
@@ -4,7 +4,7 @@
  */
 
 import * as assert from 'assert';
-import { isLogAttributeValue } from '../../../src/utils/validation';
+import { isLogAttributeValue } from '../../../src/utils/attributes';
 
 describe('isLogAttributeValue', () => {
   describe('should accept scalar values', () => {
